requirements.txt
textual>=0.54.0
dspy>=2.0.0
rich>=13.0.0,<13.7.0
pytest-timeout>=2.3.1
pytest-asyncio>=0.23.0
